+++ sdmx-dl-testing/src/main/java/sdmxdl/testing/WebRuleBatch.java
@@ -1,8 +1,0 @@
-package sdmxdl.testing;
-
-import java.util.stream.Stream;
-
-public interface WebRuleBatch {
-
-    Stream<WebRule> getProviders();
-}